# product-service/src/main/resources/openapi/product.yaml
openapi: 3.0.1
info:
  title: Product Service
  description: Product Service
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /v1.0/products:
    get:
      tags:
        - Product Controller
      summary: Get all products
      operationId: getAllProducts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllProducts'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
    post:
      tags:
        - Product Controller
      summary: Add a product
      operationId: addProduct
      requestBody:
        description: JSON object of a product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad Request response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ConstraintViolationErrorResponse'
                  - $ref: '#/components/schemas/BadRequestErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
  /v1.0/products/{id}:
    get:
      tags:
        - Product Controller
      summary: Get Product Detail
      operationId: getProductDetail
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: id of product
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'

components:
  schemas:
    GetAllProducts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        timestamp:
          type: string
          format: date-time
          example: "2024-09-08T14:32:45.123"

    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        productId:
          type: string
          example: "STD"
        effectiveFrom:
          type: string
          format: date
          example: "2024-09-08"
        effectiveTo:
          type: string
          format: date
          example: "2024-09-08"
        bookingChannel:
          type: string
          example: "IPHONE"
        productGroupId:
          type: integer
          example: 2
        isPlatformFeeApplicable:
          type: string
          example: "YES"
        remarks:
          type: string
          example: "remarks"
        platformFeeThresholdLimit:
          type: number
          format: double
          example: 1.1
        lowerPlatformFee:
          type: number
          format: double
          example: 2.3
        higherPlatformFee:
          type: number
          format: double
          example: 5
        available:
          type: boolean
          example: true
        createdBy:
          type: string
          example: "SYSTEM"
        createdDate:
          type: string
          format: date-time
          example: "2024-09-08T14:32:45.123"
        updatedBy:
          type: string
          example: "SYSTEM"
        updatedDate:
          type: string
          format: date-time
          example: "2024-09-08T14:32:45.123"

    CreateProduct:
      type: object
      properties:
        productId:
          type: string
          example: "STD"
        effectiveFrom:
          type: string
          format: date
          example: "2024-09-08"
        effectiveTo:
          type: string
          format: date
          example: "2024-09-08"
        bookingChannel:
          type: string
          example: "IPHONE"
        productGroupId:
          type: integer
          example: 2
        isPlatformFeeApplicable:
          type: string
          example: "YES"
        remarks:
          type: string
          example: "remarks"
        platformFeeThresholdLimit:
          type: number
          format: double
          example: 1.1
        lowerPlatformFee:
          type: number
          format: double
          example: 2.3
        higherPlatformFee:
          type: number
          format: double
          example: 5

    ConstraintViolationErrorResponse:
      required:
        - traceId
        - error
        - path
        - timestamp
      type: object
      properties:
        traceId:
          type: string
          example: "abcxhjvbcsjacbhbcdswmsxcasdkjcb"
        error:
          $ref: '#/components/schemas/ConstraintViolationError'
        path:
          type: string
          example: "/products"
        timestamp:
          type: string
          format: date-time
          example: "2024-09-08T14:32:45.123"

    ConstraintViolationError:
      required:
        - code
        - message
        - data
      type: object
      properties:
        code:
          type: string
          example: "ConstraintViolationError"
        message:
          type: string
          example: "Multiple errors in user request"
        data:
          $ref: '#/components/schemas/ConstraintViolationErrorData'

    ConstraintViolationErrorData:
      type: object
      properties:
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ConstraintViolationFields'

    ConstraintViolationFields:
      type: object
      properties:
        name:
          type: string
          example: "field"
        message:
          type: array
          items:
            type: string

    NotFoundErrorResponse:
      required:
        - traceId
        - error
        - path
        - timestamp
      type: object
      properties:
        traceId:
          type: string
          example: "abcxhjvbcsjacbhbcdswmsxcasdkjcb"
        error:
          $ref: '#/components/schemas/NotFoundError'
        path:
          type: string
          example: "/products"
        timestamp:
          type: string
          format: date-time
          example: "2024-09-08T14:32:45.123"

    NotFoundError:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
          example: "NotFoundException"
        message:
          type: string
          example: "Resource Not Found"

    BadRequestErrorResponse:
      required:
        - traceId
        - error
        - path
        - timestamp
      type: object
      properties:
        traceId:
          type: string
          example: "abcxhjvbcsjacbhbcdswmsxcasdkjcb"
        error:
          $ref: '#/components/schemas/BadRequestError'
        path:
          type: string
          example: "/products"
        timestamp:
          type: string
          format: date-time
          example: "2024-09-08T14:32:45.123"

    BadRequestError:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
          example: "BadRequestException"
        message:
          type: string
          example: "BadRequestError"

    InternalServerErrorResponse:
      required:
        - traceId
        - error
        - path
        - timestamp
      type: object
      properties:
        traceId:
          type: string
          example: "abcxhjvbcsjacbhbcdswmsxcasdkjcb"
        error:
          $ref: '#/components/schemas/InternalServerError'
        path:
          type: string
          example: "/products"
        timestamp:
          type: string
          format: date-time
          example: "2024-09-08T14:32:45.123"

    InternalServerError:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
          example: "InternalServerException"
        message:
          type: string
          example: "Internal Server Error"